Move branch check into the deploy script; move all env vars into circle config instead of the script

+++ .circleci/deploy.sh
@@ -1,19 +1,19 @@
 #!/usr/bin/env bash
 
-# Fail if anything within this script returns
-# a non-zero exit code
-set -e
+if [ "${CIRCLE_BRANCH}" == "master" ]
+then
+  # Fail if anything within this script returns
+  # a non-zero exit code
+  set -e
 
-# CF environment
-API="https://api.fr.cloud.gov"
-ORG="gsa-18f-hubot"
-SPACE="prod"
+  # Install CF CLI
+  curl -v -L -o cf-cli_amd64.deb 'https://cli.run.pivotal.io/stable?release=debian64&source=github'
+  sudo dpkg -i cf-cli_amd64.deb
+  rm cf-cli_amd64.deb
 
-# Install CF CLI
-curl -v -L -o cf-cli_amd64.deb 'https://cli.run.pivotal.io/stable?release=debian64&source=github'
-sudo dpkg -i cf-cli_amd64.deb
-rm cf-cli_amd64.deb
-
-# Log into CF and push
-cf login -a $API -u $CF_USERNAME -p $CF_PASSWORD -o $ORG -s $SPACE
-cf push -f /app/manifest.yml
+  # Log into CF and push
+  cf login -a $CF_API -u $CF_USERNAME -p $CF_PASSWORD -o $CF_ORG -s $CF_SPACE
+  cf push -f manifest.yml
+else
+  echo "Not on the master branch.  Not deploying."
+fi
